pyproject.toml
[tool.poetry]
name = "cfpb-complaint-triage"
version = "0.1.0"
description = "CFPB complaint triage package with training and ONNX inference."
authors = ["Петров Виктор Иванович"]
readme = "README.md"
packages = [{ include = "cfpb_complaint_triage" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fire = "^0.7.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
pydantic = "^2.10.6"
pandas = "^2.2.3"
pyarrow = "^19.0.1"
numpy = "^2.1.0"
scikit-learn = "^1.6.1"
requests = "^2.32.3"
tqdm = "^4.67.1"
torch = "^2.6.0"
lightning = "^2.5.0.post0"
transformers = "^4.49.0"
onnx = "^1.17.0"
onnxscript = "^0.6.2"
onnxruntime = "^1.20.1"
mlflow = "^2.20.1"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
dvc = "^3.59.0"
PyYAML = "^6.0.2"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.poetry.scripts]
cfpb-triage = "cfpb_complaint_triage.commands:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
